add unit test coverage for mathutil and use it in more places (#1093)
* Add MathUtil tests and fix float/double-to-int casts across poi, poi-ooxml, poi-scratchpad

* Fix long line in ImageHeaderWMF.getSize()

---------

+++ poi-scratchpad/src/main/java/org/apache/poi/hemf/record/emfplus/HemfPlusBrush.java
@@ -681,7 +681,7 @@ public class HemfPlusBrush {
             double[] end = DrawPaint.RGB2SCRGB(endColor);
 
             // compute the interpolated color in linear space
-            int a = (int)Math.round(startColor.getAlpha() + factor * (endColor.getAlpha() - startColor.getAlpha()));
+            int a = Math.toIntExact(Math.round(startColor.getAlpha() + factor * (endColor.getAlpha() - startColor.getAlpha())));
             double r = start[0] + factor * (end[0] - start[0]);
             double g = start[1] + factor * (end[1] - start[1]);
             double b = start[2] + factor * (end[2] - start[2]);
